OSINspecTor/requirements.txt
Django>=3.0,<4.0
django-crispy-forms
emoji-country-flag
ipython>=7.13.0
matplotlib>=3.4.1
requests>=2.22.0
google>=3.0.0
pandas>=1.2.4
beautifulsoup4>=4.9.3
